moving tab handling into a context
